Add Encoding and Decoding with Hex Alphabet
This patch will add encoding and decoding using the extended hex
alphabet as described in section 7 of RFC 4648.

There are two new methods with the "Hex" suffix:

  - encodeHex()
  - decodeHex()

Usage is the same as with the "normal" methods that use the standard
alphabet.

To reduce the risk of problems the test suite was extended, but may
still be lacking in terms of the test data.

Closes #13